functions/requirements.txt
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
functions-framework==3.5.0
Flask==2.3.3